move `getAllSiteDomains` to `hosting/api`
